Init implementation of middleware
