+++ system/go.mod
@@ -2,4 +2,4 @@ module system
 
 go 1.25.1
 
-require github.com/rabbitmq/amqp091-go v1.10.0 // indirect
+require github.com/rabbitmq/amqp091-go v1.10.0
