+++ system/main.go
@@ -2,10 +2,8 @@ package main
 
 import (
 	"fmt"
-
-	"system/rabbit"
 )
 
 func main() {
-    fmt.Println("hello world")
+	fmt.Println("hello world")
 }
