+++ system/middleware/queue.go
@@ -1,0 +1,6 @@
+// Where the implementation for MessageMiddlewareQueue lives
+package middleware
+
+func (q *MessageMiddlewareQueue) StartConsuming(onMessageCallback onMessageCallback) (error MessageMiddlewareError) {
+	panic("No implementada")
+}
